Adding Loot
Creating a basic entity class for the various things that will be in the dungeon.
Letting the player class entend the entity.
Creating a Loot class which is an entity.
Creating a spawn class that generates the loot in the dungeon adding this to the main file.
Recognising an interaction between the player and loot.